Abstracted out filter. TODO: implement filter functionality
